refactor: 불필요한 클래스 제거 및 패키지 구조 변경 - TextToSpeechRequest 클래스를 record로 변경하고 위치 변경 - AssistantResponse 클래스의 패키지를 model에서 service로 이동 - TextToSpeechRequest 관련 import 및 참조 수정 - 컨트롤러 메서드에 thread_id 및 scenario_id 파라미터 처리 수정
