+++ src/main/java/com/newworld/saegil/llm/service/AssistantResponse.java
@@ -1,4 +1,4 @@
-package com.newworld.saegil.llm.model;
+package com.newworld.saegil.llm.service;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
