+++ src/main/java/com/newworld/saegil/llm/service/AssistantService.java
@@ -1,6 +1,5 @@
 package com.newworld.saegil.llm.service;
 
-import com.newworld.saegil.llm.model.AssistantResponse;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface AssistantService {
