+++ src/main/java/com/newworld/saegil/llm/service/LlmProxyService.java
@@ -3,7 +3,6 @@ package com.newworld.saegil.llm.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.newworld.saegil.llm.config.ProxyProperties;
 import com.newworld.saegil.llm.config.TtsProvider;
-import com.newworld.saegil.llm.model.AssistantResponse;
 import com.newworld.saegil.simulation.service.SimulationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
